Custom Exceptions  Created

+++ inventory-service/src/main/java/az/baau/inventoryservice/service/impl/BrandServiceImpl.java
@@ -2,14 +2,17 @@ package az.baau.inventoryservice.service.impl;
 
 import az.baau.inventoryservice.dto.BrandDTO;
 import az.baau.inventoryservice.entity.Brand;
+import az.baau.inventoryservice.exception.BrandNotFoundException;
 import az.baau.inventoryservice.mapper.BrandMapper;
 import az.baau.inventoryservice.repository.BrandRepository;
 import az.baau.inventoryservice.service.BrandService;
+import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
+@Service
 public class BrandServiceImpl implements BrandService {
     private final BrandRepository brandRepository;
 
@@ -27,18 +30,28 @@ public class BrandServiceImpl implements BrandService {
 
     @Override
     public List<BrandDTO> getAllBrands() {
+
         List<Brand> brands = brandRepository.findAll();
-        List<BrandDTO> brandDTOS = new ArrayList<>();
-        for (Brand brand : brands) {
-            brandDTOS.add(BrandMapper.INSTANCE.brandToBrandDTO(brand));
+        if (brands.isEmpty() == false) {
+            List<BrandDTO> brandDTOS = new ArrayList<>();
+            for (Brand brand : brands) {
+                brandDTOS.add(BrandMapper.INSTANCE.brandToBrandDTO(brand));
+            }
+            return brandDTOS;
         }
-        return brandDTOS;
+        throw new BrandNotFoundException("Not found any brand");
+
     }
 
     @Override
     public BrandDTO getBrandById(Long id) {
         Optional<Brand> brand = brandRepository.findById(id);
-        return brand.map(BrandMapper.INSTANCE::brandToBrandDTO).orElse(null);
+        if (brand.isPresent()) {
+            return BrandMapper.INSTANCE.brandToBrandDTO(brand.get());
+        } else {
+            throw new BrandNotFoundException("id : " + id);
+        }
+
     }
 
     @Override
@@ -50,12 +63,17 @@ public class BrandServiceImpl implements BrandService {
             brand = brandRepository.save(brand);
             return BrandMapper.INSTANCE.brandToBrandDTO(brand);
         }
-        return null;
+        throw new BrandNotFoundException("Id : " + id);
     }
 
     @Override
     public void deleteBrandById(Long id) {
-        brandRepository.deleteById(id);
+        Optional<Brand> deletedBrand = brandRepository.findById(id);
+        if (deletedBrand.isPresent()) {
+            brandRepository.deleteById(id);
 
+        } else {
+            throw new BrandNotFoundException("Id : " + id);
+        }
     }
 }
