+++ inventory-service/src/main/java/az/baau/inventoryservice/service/impl/ProductServiceImpl.java
@@ -2,16 +2,21 @@ package az.baau.inventoryservice.service.impl;
 
 import az.baau.inventoryservice.dto.ProductDTO;
 import az.baau.inventoryservice.entity.Product;
+import az.baau.inventoryservice.exception.ProductNotFoundException;
 import az.baau.inventoryservice.mapper.ProductMapper;
 import az.baau.inventoryservice.repository.ProductRepository;
 import az.baau.inventoryservice.service.ProductService;
+import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
+@Service
 public class ProductServiceImpl implements ProductService {
     private final ProductRepository productRepository;
+    private final LocalDateTime localDateTime = LocalDateTime.now();
 
     public ProductServiceImpl(ProductRepository productRepository) {
         this.productRepository = productRepository;
@@ -20,6 +25,7 @@ public class ProductServiceImpl implements ProductService {
     @Override
     public ProductDTO addNewProduct(ProductDTO productDTO) {
         Product newProduct = ProductMapper.INSTANCE.productDTOtoProduct(productDTO);
+        newProduct.setRegisterDate(localDateTime);
         newProduct = productRepository.save(newProduct);
         return ProductMapper.INSTANCE.productToProductDTO(newProduct);
     }
@@ -27,9 +33,12 @@ public class ProductServiceImpl implements ProductService {
     @Override
     public List<ProductDTO> getAllProducts() {
         List<Product> products = productRepository.findAll();
-        List<ProductDTO> productDTOS = new ArrayList<>();
-        products.forEach(product -> productDTOS.add(ProductMapper.INSTANCE.productToProductDTO(product)));
-        return productDTOS;
+        if (products.isEmpty() == false) {
+            List<ProductDTO> productDTOS = new ArrayList<>();
+            products.forEach(product -> productDTOS.add(ProductMapper.INSTANCE.productToProductDTO(product)));
+            return productDTOS;
+        }
+        throw new ProductNotFoundException("Not Found Any Product");
     }
 
     @Override
@@ -39,7 +48,7 @@ public class ProductServiceImpl implements ProductService {
             Product foundProduct = optionalProduct.get();
             return ProductMapper.INSTANCE.productToProductDTO(foundProduct);
         }
-        return null;
+        throw new ProductNotFoundException("Id : " + id);
     }
 
     @Override
@@ -51,17 +60,21 @@ public class ProductServiceImpl implements ProductService {
             updatedProduct.setPrice(productDTO.getPrice());
             updatedProduct.setDescription(productDTO.getDescription());
             updatedProduct.setQuantity(productDTO.getQuantity());
-            updatedProduct.setUpdateDate(productDTO.getUpdateDate());
+            updatedProduct.setUpdateDate(localDateTime);
             productRepository.save(updatedProduct);
             return ProductMapper.INSTANCE.productToProductDTO(updatedProduct);
         }
-        return null;
+        throw new ProductNotFoundException("Id : " + id);
 
     }
 
     @Override
     public void deleteProductById(Long id) {
-        productRepository.deleteById(id);
-
+        Optional<Product> deletedProduct = productRepository.findById(id);
+        if (deletedProduct.isPresent()) {
+            productRepository.deleteById(id);
+        } else {
+            throw new ProductNotFoundException("Id : " + id);
+        }
     }
 }
